Refactor print_test to print_example
