+++ code/svb-docker/python/svb_web/core/views/__init__.py
@@ -7,4 +7,4 @@ from .internal.common import *
 from .internal.banker import *
 from .internal.display import *
 from .internal.print_receipt import *
-from .internal.print_test import *
+from .internal.print_example import *
