+++ code/svb-docker/python/svb_web/core/views/internal/print_example.py
